+++ app/src/main/java/com/hexabinome/saladetomateoignon/modele/Restaurant.java
@@ -10,16 +10,16 @@ import java.util.List;
 public class Restaurant {
     //TODO: ADD PHOTOS OF RESTAURANTS
     private String name;
-    private Double prix;
-    private Double tempsAttenteMoy;
-    private Double distance; // TODO à remplacer par position, puis calculer la distance
-    private Integer note;
+    private double prix;
+    private int tempsAttenteMoy;
+    private double distance; // TODO à remplacer par position, puis calculer la distance
+    private double note;
     private TypePointDeRestauration typePointDeRestauration;
     private List<TypeRegime> regimeList;
 
 
-    public Restaurant(String myname, double myprice, double tempsAttenteMoy, double distance,
-                      int note, TypePointDeRestauration typePointDeRestauration, List<TypeRegime> regimes) {
+    public Restaurant(String myname, double myprice, int tempsAttenteMoy, double distance,
+                      double note, TypePointDeRestauration typePointDeRestauration, List<TypeRegime> regimes) {
         this.prix = myprice;
         this.name = myname;
         this.tempsAttenteMoy = tempsAttenteMoy;
@@ -33,19 +33,19 @@ public class Restaurant {
         return name;
     }
 
-    public Double getPrix() {
+    public double getPrix() {
         return prix;
     }
 
-    public Double getTempsAttenteMoy() {
+    public int getTempsAttenteMoy() {
         return tempsAttenteMoy;
     }
 
-    public Double getDistance() {
+    public double getDistance() {
         return distance;
     }
 
-    public Integer getNote() {
+    public double getNote() {
         return note;
     }
